requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
openai>=1.0.0
anthropic>=0.30.0
PyGithub>=2.3.0
gitpython>=3.1.40
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
black>=24.0.0
ruff>=0.4.0
mypy>=1.10.0
